+++ integration_tests/suite/test_phone_view.py
@@ -16,9 +16,3 @@ class TestPhoneView(BaseDirdIntegrationTest):
     asset = 'phone_view'
     profile = 'default'
     config_factory = new_phone_view_config
-
-    def test_thomson_lookup_route(self):
-        result = self.get_lookup_thomson_result(
-            self.profile, VALID_UUID, term='Alice', token=VALID_TOKEN_MAIN_TENANT
-        )
-        assert_that(result.status_code, equal_to((200)))